target/maltlab/requirements-baremetal.txt
grpcio
pyyaml
docopt
numpy
requests
#smartredis==0.3.1
#smartredis==0.5.0
tensorflow-serving-api
cloudmesh-common==4.4.1
cloudmesh-cmd5==4.3.33
cloudmesh-rivanna==4.3.21

tqdm
smartredis
matplotlib
#cloudmesh-ee
#cloudmesh-gpu

# git+https://github.com/mlperf/logging.git@2.1.0
# git+https://github.com/mlperf/logging.git@1.0.0
